% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{bplot}
\alias{bplot}
\title{Plot bargraph}
\usage{
bplot(datset, x, y, group = NULL, e, xlab = NULL, ylab = NULL)
}
\arguments{
\item{x}{The independent variable to be plotted on the x axis. Must be categorical.}

\item{y}{The dependent variable to be plotted on the y axis.}

\item{group}{A second grouping variable.}

\item{xlab}{A title for the x-axis. If not defined, will return name of x variable.}

\item{ylab}{A title for the y-axis. If not defined, will return name of y variable.}

\item{data}{A dataframe}

\item{se}{The standard error variable to create error bars.}
}
\description{
This function will plot means in a bargraph for for 1 or 2 categorical variables.
}
\examples{
aa <- data.frame(condition = c("neutral","positive","negative"),
                accuracy = c(.6, .74, .8),
                se_accuracy = c(.04, .08, .09))
bplot(datset = aa, x = "condition", y = "accuracy", e = "se_accuracy",
        xlab = "Experimental Condition", ylab = "Mean Accuracy")
}
\keyword{barplot}
