% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{corr_cvplot}
\alias{corr_cvplot}
\title{Plot Corr-Val}
\usage{
corr_cvplot(x, y, xlab = NULL, ylab = NULL)
}
\arguments{
\item{x}{A single vector.}

\item{y}{A single vector.}
}
\value{
A plot of the correlation between x and y, with the cross-validated correlation coefficient annotated within the graph.
}
\description{
The function will plot a correlation with an annotation of the cross-validated pearson coefficient.
}
\examples{
x <- rnorm(n = 25, mean = 0, sd = 1)
y <- rnorm(n = 25, mean = 25, sd = 5)
corr_cv(x, y)
}
\keyword{correlation}
\keyword{cross-validated}
