% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dime}
\alias{dime}
\title{DIME Connectivity Data}
\format{A dataframe with 24 rows of 47 variables:}
\source{
Brain and Memory Lab
}
\usage{
dime
}
\description{
This dataset contains the rest and residual connectivity from the DIME study. It includes connectivity data
from pre-encoding rest, encoding scans averaged, and post-encoding rest. Additionally, this dataframe has relevant
behavioral measures.
}
\keyword{datasets}
