% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{se}
\alias{se}
\title{Standard Error}
\usage{
se(x)
}
\arguments{
\item{x}{A single vector.}
}
\description{
This function allows you to calculate standard error
}
\examples{
se(x)
}
\keyword{error}
\keyword{standard}
